requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0

pydantic>=2.5.0
python-dotenv>=1.0.0
loguru>=0.7.0

langchain>=0.1.0
langchain-ollama>=0.1.0
langchain-community>=0.0.10
ollama>=0.1.6

sentence-transformers>=2.3.1
faiss-cpu>=1.7.4

pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
